* Reworked route editing backend * rewireing is done by RouteClass/GUIRoute now * other classes forward most requests to this pair * many bugfixes * Version change: We are reaching beta phase now

+++ mu32/routing/gmn/GSP.cpp
@@ -2,16 +2,23 @@
  ********************************************************************
  * Description
  *
- * $Header: /home/tobias/macbookbackup/Entwicklung/mutabor/cvs-backup/mutabor/mutabor/mu32/routing/gmn/GSP.cpp,v 1.4 2011/02/20 22:35:56 keinstein Exp $
+ * $Header: /home/tobias/macbookbackup/Entwicklung/mutabor/cvs-backup/mutabor/mutabor/mu32/routing/gmn/GSP.cpp,v 1.5 2011/09/27 20:13:22 keinstein Exp $
  * Copyright:   (c) 2008 TU Dresden
  * \author  Tobias Schlemmer <keinstein@users.berlios.de>
  * \date 
- * $Date: 2011/02/20 22:35:56 $
- * \version $Revision: 1.4 $
+ * $Date: 2011/09/27 20:13:22 $
+ * \version $Revision: 1.5 $
  * \license GPL
  *
  * $Log: GSP.cpp,v $
- * Revision 1.4  2011/02/20 22:35:56  keinstein
+ * Revision 1.5  2011/09/27 20:13:22  keinstein
+ * * Reworked route editing backend
+ * * rewireing is done by RouteClass/GUIRoute now
+ * * other classes forward most requests to this pair
+ * * many bugfixes
+ * * Version change: We are reaching beta phase now
+ *
+ * Revision 1.4  2011-02-20 22:35:56  keinstein
  * updated license information; some file headers have to be revised, though
  *
  *
@@ -538,7 +545,7 @@ int ReadNote()
 	if ( SepPos )
 		return CheckError(Note(Name, accedentials, octave, duration));
 
-	int cp = CurrentPos; // octave
+	size_t cp = CurrentPos; // octave
 
 	int oct = ReadLong(1);
